fastapi_backend/requirements.txt
# FastAPI
fastapi>=0.103.1
uvicorn>=0.23.2

# Database
sqlalchemy>=2.0.20
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Data validation
pydantic>=2.3.0
pydantic-settings>=2.0.3
email-validator>=2.0.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Environment variables
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.27.0

# AI/ML
groq>=0.4.1

# Email
aiosmtplib>=2.0.1

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1 